# gateway/cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/category/create": {
            "post": {
                "description": "Добавляет новую категорию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Создать категорию",
                "parameters": [
                    {
                        "description": "Новая категория",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно создана",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/category/delete": {
            "delete": {
                "description": "Удаляет категорию по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Удалить категорию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID категории",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/category/getall": {
            "get": {
                "description": "Возвращает список всех категорий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Получить все категории",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.SWGCategoryListResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/category/update": {
            "put": {
                "description": "Обновляет информацию о категории",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Обновить категорию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID категории",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Обновлённая категория",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно обновлена",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или формат",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "views.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGCategoryListResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Category"
                    }
                }
            }
        },
...
        "views.SWGErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "views.SWGSuccessResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        }
    }
}